refactor(Structure): Product View(상품 상세 정보 뷰) 구조 리팩토링 및 네트워크 통신 기능 추가
- urlsession을 활용한 이미지 로딩 진행
- viewController로부터 View 분리
- viewModel Binding
