+++ src/ios/Cream/Data/Network/APIEndpoints.swift
@@ -9,6 +9,7 @@ import Foundation
 
 struct APIEndpoints {
 
+    // MARK: - User
     static func confirmUser(with authRequestDTO: AuthRequestDTO) -> Endpoint<AuthResponseDTO> {
         return Endpoint(path: "users/login",
                         method: .post,
@@ -21,6 +22,14 @@ struct APIEndpoints {
                         method: .post,
                         headerParamaters: ["Content-Type":"application/json"],
                         bodyParamatersEncodable: joinRequestDTO)
+    }
+    
+    // MARK: - Product
+    static func loadProduct(_ id: Int) -> Endpoint<ProductResponseDTO> {
+        // TODO: AccessToken이 있다면, header에 해당 값 넣기
+        return Endpoint(path: "products/\(id)",
+                        method: .get,
+                        headerParamaters: ["Content-Type":"application/json"])
     }
     
 }
