+++ src/ios/Cream/Presentation/ProductScene/ProductViewModel.swift
@@ -2,67 +2,75 @@
 //  ProductViewModel.swift
 //  Cream
 //
-//  Created by wankikim-MN on 2022/01/24.
+//  Created by wankikim-MN on 2022/01/30.
 //
 
 import Foundation
 
-enum Category: String, CaseIterable {
-    case slide = "slid"
-    case luxury = "럭셔리"
-    case blank = " "
-    case sneakers = "스니커즈"
-    case clothes = "의류"
-    case accessory = "패션 잡화"
-    case life = "라이프"
-    case tech = "테크"
+
+protocol ProductViewModelInput {
+    func viewDidLoad()
+    func didTapWishButton()
+    func didTapBuyButton()
+    func didTapSellButton()
 }
 
-class ProductsViewModel: NSObject {
-    private var productService: ProductsServiceProtocol
-    var reloadCollectionView: (() -> Void)?
-//    let categories = Category.allCases
-    let categories = ["필터", "럭셔리", " ", "스니커즈", "의류", "패션 잡화", "라이프", "테크"]
-    let banners: [String] = ["banner1", "banner2", "banner3", "banner4", "banner5", "banner6"]
+protocol ProductViewModelOutput {
+    var item: Observable<ProductDetail> { get set }
+    var count: Int { get }
+    var releaseInfo: [(String, String)] { get }
+}
+
+protocol ProductViewModel: ProductViewModelInput, ProductViewModelOutput {
+    var usecase: ProductUseCaseInterface { get set }
+}
+
+final class DefaultProductViewModel: ProductViewModel {
+    var usecase: ProductUseCaseInterface
     
-    var products: Products = Products()
+    var item: Observable<ProductDetail> = Observable(ProductDetail.create())
+
+    let imageUrls: [String] = ["mock_shoe1", "mock_shoe2", "mock_shoe3"]
     
-    var productCellViewModels = [ProductCellViewModel]() {
-        didSet {
-            self.reloadCollectionView?()
-        }
+    var releaseInfo: [(String, String)] {
+        var info = [(String, String)]()
+        info.append(("모델 번호", item.value.styleCode))
+        info.append(("출시일", item.value.releaseDate))
+        info.append(("대표색상", item.value.color))
+        info.append(("발매가", item.value.originalPrice.priceFormat))
+        
+        return info
+    }
+    private var page: Int = 1
+    
+    var count: Int {
+        return imageUrls.count
     }
     
-    var productCellViewModel: Observable<[ProductCellViewModel]> = Observable([])
-    
-    init(productService: ProductsServiceProtocol = ProductsService()) {
-        self.productService = productService
+    init(usecase: ProductUseCaseInterface) {
+        self.usecase = usecase
     }
     
-    func getProducts() {
-        productService.getProducts { (success, results, error) in
-            if success, let products = results {
-                self.fetchData(products: products)
+    func viewDidLoad() {
+        usecase.fetchItemById(2351) { result in
+            switch result {
+            case .success(let product):
+                self.item.value = product
+            case .failure(let error):
+                print(error)
             }
         }
     }
     
-    func fetchData(products: Products) {
-        self.products = products // Cache
-        var viewModels = [ProductCellViewModel]()
-        for item in products {
-            viewModels.append(createCellModel(product: item))
-        }
+    func didTapWishButton() {
+         
     }
     
-    func createCellModel(product: Product) -> ProductCellViewModel {
-        return ProductCellViewModel(totalSale: "\(product.productInfo.totalSale)",
-                                    brandName: product.productInfo.brandName,
-                                    originalName: product.productInfo.originalName,
-                                    backgroundColor: product.productInfo.backgroundColor,
-                                    imageUrls: product.productInfo.imageUrls,
-                                    wishList: product.wishList,
-                                    lowestAsk: "\(product.lowestAsk)",
-                                    wishCount: "\(product.productInfo.wishCount)")
+    func didTapBuyButton() {
+         
+    }
+    
+    func didTapSellButton() {
+         
     }
 }
