+++ src/ios/Cream/Presentation/ProductScene/ProductViewModelItem.swift
@@ -46,7 +46,6 @@ class ProductViewModelImageItem: ProductViewModelItem {
     }
 }
 
-
 struct Info {
     var brand: String
     var detailInfo: String
